update trex profiles
