+++ trex_profiles/http_max_cps.py
@@ -11,12 +11,13 @@ http_req = 'GET /index.html HTTP/1.1\r\n' \
            'Accept-Encoding: gzip, deflate, compress\r\n' \
            '\r\n'
 
-http_response_template = 'HTTP/1.1 200 OK\r\n' \
+http_response_headers = 'HTTP/1.1 200 OK\r\n' \
                          'Server: Microsoft-IIS/6.0\r\n' \
                          'Content-Type: text/html\r\n' \
                          'Content-Length: {0}\r\n' \
-                         '\r\n' \
-                         '<html><pre>{1}</pre></html>'
+                         '\r\n'
+
+http_response_body = '<html><pre>{0}</pre></html>'
 
 RESPONSE_BODY_TAGS_BYTES = 24
 
@@ -25,9 +26,12 @@ class Prof1():
     def __init__(self):
         pass  # tunables
 
-    def create_profile(self, resp_size):
-
-        http_response = http_response_template.format(resp_size, '*' * ( resp_size - RESPONSE_BODY_TAGS_BYTES ))
+    def create_profile(self, resp_body_size, delay, ramp_up):
+        if resp_body_size > 24:
+            http_response = http_response_headers.format(resp_body_size) + \
+                            http_response_body.format('*' * (resp_body_size - RESPONSE_BODY_TAGS_BYTES))
+        else:
+            http_response = http_response_headers.format(resp_body_size) + '*' * resp_body_size
         # client commands
         prog_c = ASTFProgram()
         prog_c.send(http_req)
@@ -37,6 +41,7 @@ class Prof1():
         # server commands
         prog_s = ASTFProgram()
         prog_s.recv(len(http_req))
+        if delay: prog_s.delay(delay)
         prog_s.send(http_response)
         prog_s.wait_for_peer_close()
 
@@ -56,6 +61,7 @@ class Prof1():
         c_glob_info = ASTFGlobalInfo()
         c_glob_info.tcp.rxbufsize = 65535
         c_glob_info.tcp.txbufsize = 65535
+        if ramp_up: c_glob_info.scheduler.rampup_sec = ramp_up
         s_glob_info = ASTFGlobalInfo()
         s_glob_info.tcp.rxbufsize = 65535
         s_glob_info.tcp.txbufsize = 65535
@@ -70,15 +76,23 @@ class Prof1():
     def get_profile(self, tunables, **kwargs):
         parser = argparse.ArgumentParser(description='Argparser for {}'.format(os.path.basename(__file__)), 
                                          formatter_class=argparse.ArgumentDefaultsHelpFormatter)
-        parser.add_argument('--size',
-                            type=int,
+        parser.add_argument('--response_size_kb',
+                            type=float,
                             default=1,
                             help='The response size in KB')
+        parser.add_argument('--delay',
+                            type=int,
+                            default=None,
+                            help='Server response delay (usec)')
+        parser.add_argument('--ramp_up',
+                            type=int,
+                            default=0,
+                            help='Ramp up period (sec)')
         args, unknown = parser.parse_known_args(tunables)
         if unknown:
             raise Exception('unrecognized arguments {0}\n{1}'.format(unknown, parser.format_usage()))
-        resp_size = args.size * 1024
-        return self.create_profile(resp_size)
+        resp_size = int(args.response_size_kb * 1024)
+        return self.create_profile(resp_size, args.delay, args.ramp_up)
 
 
 def register():
